Post-Jam Fixes
Fixed camera bug where it would be outside of the viewport when the window size is increased (e.g. fullscreen mode).

Expanded the levels' ColorRect to the left one pixel.

Enabled looping on the music.

Added tiles to the edges of the TileMap. Also added tiles to prevent exploring empty areas.